auth and register

+++ notes.controller.js
@@ -1,8 +1,8 @@
 const chalk = require("chalk");
 const Note = require("./models/note");
 
-async function addNote(title) {
-  await Note.create({ title });
+async function addNote(title, owner) {
+  await Note.create({ title, owner });
   console.log(chalk.bgGreen("Note was added"));
 }
 
@@ -12,14 +12,25 @@ async function getNotes() {
   return notes;
 }
 
-async function removeNoteById(id) {
-  await Note.deleteOne({ _id: id });
+async function removeNoteById(id, owner) {
+  const result = await Note.deleteOne({ _id: id, owner });
+
+  if (result.matchedCount === 0) {
+    throw new Error("No note to delete");
+  }
 
   console.log(chalk.bgRed(`Note with id ${id} was removed`));
 }
 
-async function updateNoteById(noteData) {
-  await Note.updateOne({ _id: noteData.id }, { title: noteData.title });
+async function updateNoteById(noteData, owner) {
+  const result = await Note.updateOne(
+    { _id: noteData.id, owner },
+    { title: noteData.title }
+  );
+
+  if (result.matchedCount === 0) {
+    throw new Error("No note to edit");
+  }
   console.log(
     chalk.bgGreen(
       `Note with id ${noteData.id} was updated, new title - ${noteData.title}`
